deep_learning/requirements.txt
# 核心框架
torch>=2.0.0
torchaudio>=2.0.0

# 数据处理
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 配置管理
pyyaml>=6.0

# ONNX部署
onnx>=1.14.0
onnxruntime>=1.15.0

# 可视化(可选)
matplotlib>=3.5.0

# 进度条
tqdm>=4.64.0
